Relocate ExportFormat interface to model
