+++ web-app/src/app/locales/model/exportformat.ts
@@ -1,0 +1,5 @@
+export interface ExportFormat {
+    name: string;
+    extension: string;
+    apiIdent: string;
+}
