+++ web-app/src/app/locales/model/index.ts
@@ -1,2 +1,3 @@
 export * from './locale';
 export * from './localeinfo';
+export * from './exportformat';
